fix search customer_id fields int size

+++ app/Stubs/Migrations/HasSearchFields.php
@@ -51,7 +51,7 @@ trait HasSearchFields
         }
 
 
-        $table->unsignedSmallInteger('customer_id')->nullable()->index();
+        $table->unsignedInteger('customer_id')->nullable()->index();
         $table->foreign('customer_id')->references('id')->on('customers');
 
         return $this->searchFields($table);
